refactor: changed the file name of run_statistics
