+++ src/lib.rs
@@ -1,1 +1,1 @@
-pub mod function_benchmark;
+pub mod run_statistics;
